@mshabunin 2025-01-25
I've tested "updated code" (after) VS "current 4.x" (new) VS "4.9" (old). Also profiled with callgrind to check number of memory allocations. Each measurment took 50 iterations using reproducer from the original issue.
|Name of Test|before|new|after|new vs before (x-factor)|after vs before (x-factor)|
|---|:-:|:-:|:-:|:-:|:-:|
|findContours_big::TestFindContours|41.837|75.655|61.264|0.55|0.68|
| malloc calls | 200K | 23M | 13M | | |

Profiling with callgrind shows that updated code still have numerous malloc calls. I've been able to reduce this number even more (by ~3M) by disabling `index_mapping` access in the `cv::contourTreeToResults` fuinction and get performance ~45ms - comparable with old variant.