@chacha21 2025-01-25
> Perhaps we should use some storage with blocks allocation in your arena, something similar to old C-API `MemStorage`. Thus we will avoid vector reallocations which happens currently.

`contourTreeToResults()` is called once, so there is certainly no improvement by pre-allocating the memory needed for index_mapping. A stack allocation would help, but the current test case has a `tree.size()` equal to 127384 elements, which is too large for that.