@mshabunin 2025-01-26
> `contourTreeToResults()` is called once, so there is certainly no improvement by pre-allocating the memory needed for index_mapping

Did you measure performance difference with `index_mapping` update on your platform? For me it gave 60 ms -> 45 ms improvement on Linux/GCC.

> Perhaps we should use some storage with blocks allocation in your arena, something similar to old C-API MemStorage. Thus we will avoid vector reallocations which happens currently.

This comment was related to points stored in contours, not to TreeIterator. I believe TreeIterator doesn't have big impact on performance.