@mshabunin 2025-01-26
Please take a look at this commit: https://github.com/opencv/opencv/commit/5f35ea3dfa71a52e3c4ff4d9d3871e9cc812836b
This is block storage which I was writing about. Though I'm not sure how to make it more concise. 
On my machine it works for 42 ms VS 40 ms for legacy function. Number of malloc calls is 275K VS 185K. There is still some room for improvement (for example, perhaps we dont need to store reference to storage in each contour, but add is an argument in corresponing methods), but overall it look much better.