@chacha21 2025-01-26
> Please take a look at this commit: [5f35ea3](https://github.com/opencv/opencv/commit/5f35ea3dfa71a52e3c4ff4d9d3871e9cc812836b)

What I understand :

- this commit is about replacing the Arena (which has a policy as fixed size, then dynamic allocation when static capacity is reached) by a structure performing incremental dynamic allocations by blocks.
- I can see that you tested a static size of 600000 bytes for the arena, because of the current test case generating a high number of points. Of course, it would not be a good idea to allocate that much memory for the standard case. Hence it seems doomed, and the Block storage might be better in general.
- the Block storage is not conceptually very different from a vector. It performs dynamic allocations as it grows.
- the difference is that a vector will not rely on realloc(), it will always copy all its previous content, while the block storage will just add new blocks
- your `BlockStorage` is similar to the `vectorWithArena` that I tried previously for `TreeIterator::levels` (but vectorWithArena allocated blocks from an arena instead of the standard `new`)
- for the current test case, requiring a high amount of memory, a small arena with a fixed stack-allocated buffer won't help, but for the general case it could still be useful for the general case.

A will test the performance with your commit on my host machine ASAP.
I think that I will then evolve to a BlockStorageWithArena for the general case and get rid of the not so helpful TreeNodes arena. It will be a first cleanup.