@mshabunin 2025-01-27
As I understand this algorithm requires memory storage with :
1. O(1) push_back
2. O(1) random access 
3. elements for each contour are added sequentially
4. minimum number of allocations and deallocations

- std::vector satisfies 1-3, but not 4 - due to reallocations.
- std::deque should work like this (1-3), but it seems that it uses [quite small blocks](https://devblogs.microsoft.com/oldnewthing/20230810-00/?p=108587) (especially in MSVC - 16 bytes or 1 element), thus 4 is not satisfied as well

I thought about using std::deque with custom allocator and larger blocks, but it seems to make things too complicated.
Do we really need arena in this case? Wouldn't it add extra overhead for keeping items?

>  I am not sure that contour points should be stored as int2(x,y), but rather int(offset)

Not sure if I understand. Wouldn't offset need to be 2D?