@chacha21 2025-01-27
> As I understand this algorithm requires memory storage with : [...]
>     * std::vector satisfies 1-3, but not 4 - due to reallocations.
>     * std::deque should work like this (1-3), but it seems that it uses [quite small blocks](https://devblogs.microsoft.com/oldnewthing/20230810-00/?p=108587) (especially in MSVC - 16 bytes or 1 element), thus 4 is not satisfied as well
> I thought about using std::deque with custom allocator and larger blocks, but it seems to make things too complicated. 

I think you are right. We should not rely on a structure which inner tuning depends heavily on the toolchain. Your BlockStorage let us control the amount of memory reliably.

>Do we really need arena in this case? Wouldn't it add extra overhead for keeping items?

The idea of the arena is just to provide fast memory on demand, I don't think that my ArenaStackBuffer has much overhead.
However, even better, I think that you could just tune your BlockStorage with one static buffer holding the first blocks (that will be allocated on stack if the BlockStorage itself is allocated on stack before creating the ContourScanner). When more blocks are needed, the dynamic allocation will occur.
No need to delegate that to an arena. 


> > I am not sure that contour points should be stored as int2(x,y), but rather int(offset)
> Not sure if I understand. Wouldn't offset need to be 2D?

Yes, but as long as we keep track of width, offset = y*width+x, so (x, y) = (offset%width, offset/width). This is just a trick to save half the memory used to store one point (at the cost of the offset <-> (x,y) conversion overhead here and there)